Total restructure of codebase and addition of 'whitelist' as required parameter for account input
